# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/housing-units/{id}": {
      "get": {
        "operationId": "HousingUnitsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingUnitDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a housing unit you own or belong to",
        "tags": [
          "housing-units"
        ]
      },
      "patch": {
        "operationId": "HousingUnitsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHousingUnitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingUnitDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a housing unit (owner only)",
        "tags": [
          "housing-units"
        ]
      },
      "delete": {
        "operationId": "HousingUnitsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a housing unit (owner only)",
        "tags": [
          "housing-units"
        ]
      }
    },
...
    "/housing-units/{unitId}/invitations": {
      "post": {
        "operationId": "UnitInvitationsController_invite",
        "parameters": [
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invite someone to the unit by email (owner only)",
        "tags": [
          "invitations"
        ]
      },
      "get": {
        "operationId": "UnitInvitationsController_list",
        "parameters": [
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List invitations for a unit (owner only)",
        "tags": [
          "invitations"
        ]
      }
    },
    "/housing-units/{unitId}/invitations/{invitationId}": {
      "delete": {
        "operationId": "UnitInvitationsController_revoke",
        "parameters": [
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke an invitation (owner only)",
        "tags": [
          "invitations"
        ]
      }
    },
    "/housing-units/{unitId}/members": {
      "get": {
        "operationId": "UnitInvitationsController_members",
        "parameters": [
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List members of a unit",
        "tags": [
          "invitations"
        ]
      }
    },
    "/housing-units/{unitId}/members/{memberUserId}": {
      "delete": {
        "operationId": "UnitInvitationsController_removeMember",
        "parameters": [
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a member from a unit (owner only)",
        "tags": [
          "invitations"
        ]
      }
    },
    "/invitations/lookup": {
      "get": {
        "operationId": "InvitationsController_lookup",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationLookupDto"
                }
              }
            }
          }
        },
        "summary": "Look up an invitation by token (public)",
        "tags": [
          "invitations"
        ]
      }
    },
    "/invitations/accept": {
      "post": {
        "operationId": "InvitationsController_accept",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accept an invitation (must be signed in as the invited email)",
        "tags": [
          "invitations"
        ]
      }
    },
    "/invitations/reject": {
      "post": {
        "operationId": "InvitationsController_reject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reject an invitation",
        "tags": [
          "invitations"
        ]
      }
    },
    "/rooms/{roomId}/reservations": {
      "post": {
        "operationId": "RoomReservationsController_create",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reserve a room (validates room rules and overlaps)",
        "tags": [
          "reservations"
        ]
      },
      "get": {
        "operationId": "RoomReservationsController_list",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservationDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Room availability (tenants see anonymized slots)",
        "tags": [
          "reservations"
        ]
      }
    },
...
    "/reservations/mine": {
      "get": {
        "operationId": "ReservationsController_mine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MyReservationDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List my reservations across all units",
        "tags": [
          "reservations"
        ]
      }
    },
    "/reservations/{id}": {
      "patch": {
        "operationId": "ReservationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update your reservation (re-validates rules)",
        "tags": [
          "reservations"
        ]
      }
    },
    "/reservations/{id}/cancel": {
      "post": {
        "operationId": "ReservationsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel your reservation",
        "tags": [
          "reservations"
        ]
      }
    }
  },
  "info": {
    "title": "RoomMate API",
    "description": "Shared-housing room reservation API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "HousingUnitDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "viewerRole": {
            "type": "string",
            "enum": [
              "OWNER",
              "MEMBER"
            ]
          },
          "roomCount": {
            "type": "integer"
          },
          "memberCount": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "description",
          "ownerId",
          "createdAt",
          "updatedAt",
          "viewerRole",
          "roomCount",
          "memberCount"
        ]
      },
      "UpdateHousingUnitDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "address": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          }
        }
      },
...
      "CreateInvitationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          }
        },
        "required": [
          "email"
        ]
      },
      "InvitationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "unitId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REJECTED",
              "EXPIRED"
            ]
          },
          "token": {
            "type": "string"
          },
          "sentAt": {
            "type": "string",
            "format": "date-time"
          },
          "respondedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "unitId",
          "email",
          "status",
          "token",
          "sentAt",
          "respondedAt"
        ]
      },
      "MemberDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "userId",
          "email",
          "firstName",
          "lastName",
          "joinedAt"
        ]
      },
      "InvitationLookupDto": {
        "type": "object",
        "properties": {
          "unitName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REJECTED",
              "EXPIRED"
            ]
          }
        },
        "required": [
          "unitName",
          "email",
          "status"
        ]
      },
      "TokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          }
        },
        "required": [
          "token"
        ]
      },
      "CreateReservationDto": {
        "type": "object",
        "properties": {
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": "string",
            "format": "date-time"
          },
          "note": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          }
        },
        "required": [
          "startAt",
          "endAt"
        ]
      },
      "ReservationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "roomId": {
            "type": "string",
            "format": "uuid"
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": "string",
            "format": "date-time"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "CANCELLED"
            ]
          },
          "isMine": {
            "type": "boolean"
          },
          "author": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "firstName",
              "lastName"
            ],
            "nullable": true
          }
        },
        "required": [
          "id",
          "roomId",
          "startAt",
          "endAt",
          "note",
          "status",
          "isMine",
          "author"
        ]
      },
      "MyReservationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "roomId": {
            "type": "string",
            "format": "uuid"
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": "string",
            "format": "date-time"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "CANCELLED"
            ]
          },
          "isMine": {
            "type": "boolean"
          },
          "author": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "firstName",
              "lastName"
            ],
            "nullable": true
          },
          "roomName": {
            "type": "string"
          },
          "unitId": {
            "type": "string",
            "format": "uuid"
          },
          "unitName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "roomId",
          "startAt",
          "endAt",
          "note",
          "status",
          "isMine",
          "author",
          "roomName",
          "unitId",
          "unitName"
        ]
      },
      "UpdateReservationDto": {
        "type": "object",
        "properties": {
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": "string",
            "format": "date-time"
          },
          "note": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          }
        }
      }
    }
  }
}